Add ui for selective disclosure confimration
