+++ src/ux/step2/index.tsx
@@ -2,7 +2,6 @@
 import React, { useState } from 'react'
 import { SelectiveDisclosureRequest, SDR } from './SelectiveDisclosureRequest'
 import { SelectiveDisclosureResponse, SD, Data } from './SelectiveDisclosureResponse'
-import { Header2 } from '../../ui/shared/Typography'
 
 interface Step2Props {
   sdr: {
@@ -29,7 +28,6 @@ const SelectiveDisclosure = ({ sdr, backendUrl, fetchSelectiveDisclosureRequest,
   }
 
   return <>
-    <Header2>Would you like to give us<br />access to info in your data vault?</Header2>
     {!sdrConfirmed
       ? <SelectiveDisclosureRequest sdr={sdr} backendUrl={backendUrl} onConfirm={onSdrConfirm} />
       : <SelectiveDisclosureResponse data={data} backendUrl={backendUrl} onConfirm={onConfirm} />}
